devops/proyecto_asir/auth-service/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
email-validator==2.1.0.post1
disposable-email-domains==0.0.98
requests==2.31.0
itsdangerous==2.1.2
gunicorn==21.2.0
